Special case to read all fields of dns records from the api
In order to hold on to fields only available in more specific
dns record types on api calls which return the base record type,
we need to add the record-type specific fields to the base record type.
Composition will not help here since it is not possible to parse in
a json array containing different types of structs (i.e. different
types of dns records) using a base struct that does not have all the
necessary fields up front.