Do not abbreviate method name for createObjects REST calls
The client was assuming that one or the other, createObject or createObjects,
would exist as service methods, but not both. So it omitted the method name
in both of those cases expecting the api to default to the create method
that exists internally. Here we have a case where a service has two create
methods, one singular and one plural, and the default one is the singular one
when you omit the method name from the REST url. Thus a user would never
be able to use the plural create method through the client.

Closes #76